RC #251 - Combining LocationMarker and PlaceMarker components; Removing @peripleo/maplibre dependency from geospatial
